backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

torch>=2.4.0
torchaudio>=2.4.0
torchvision>=0.19.0

openai-whisper>=20231117
chatterbox-tts>=0.1.0

numpy>=1.24.0,<2.0.0
av>=13.0.0
ffmpeg-python>=0.2.0
librosa>=0.10.0

aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
websockets>=12.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
